More moving of code into logical places
This is work towards #803

Finally, the code is organized in this way:

default <-- all default operation code lives here
   |_container <-- all container image analysis functions live here
         |_run.py <-- container analysis entry point
         |_image.py <-- container image analysis
         |_single_layer.py <-- analysis of 1 layer
         |_multi_layer.py <-- analysis of more than 1 layer using
                              a copy-on-write filesystem driver
   |_dockerfile <-- all Dockerfile analysis functions live here
         |_run.py <-- Dockerfile analysis entry point
   |_core.py <-- core default execution functions
   |_collect.py <-- process results of invoked commands
   |_filter.py <-- modify Command objects
   |_bundle.py <-- convert the results into Package, FileData,
                   Notice, NoticeOrigin and Origins objects to
                   add to the ImageLayer object

core.py is the default execution "special sauce" and we usually
don't mess with it. The other functions should be free game.

The overall code execution is still broken.
