+++ tern/analyze/default/collect.py
