+++ tern/analyze/default/filter.py
